Fix can_be_removed flag not works on placed plates and spoons

+++ src/main/java/com/github/argon4w/hotpot/placements/HotpotPlacedSpoon.java
@@ -44,7 +44,9 @@ public class HotpotPlacedSpoon implements IHotpotPlacement {
             int layer,
             LevelBlockPos pos,
             IHotpotPlacementContainer container) {
-        onRemove(container, pos);
+        if (container.canBeRemoved()) {
+            onRemove(container, pos);
+        }
     }
 
     @Override
